worker/requirements.txt
# Worker service dependencies
# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.4
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and event streaming
redis==5.0.1
aioredis==2.0.1

# AI providers
openai>=1.50.0
langchain>=0.3.0
langchain-gigachat==0.3.12
langchain-core>=0.3.0

# Vector database
qdrant-client==1.8.0

# Graph database
neo4j==5.15.0

# Web crawling and enrichment
crawl4ai==0.7.6
beautifulsoup4==4.12.2
lxml>=5.3.0
playwright>=1.49.0

# HTTP client
httpx>=0.27.2
aiohttp>=3.11.11

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging and debugging
rich>=13.9.4
loguru==0.7.2